backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.8.0
pydantic-settings>=2.3.0
httpx>=0.27.0
openai>=1.30.0
chromadb>=0.5.0
python-dotenv>=1.0.1
prisma>=0.11.0
boto3>=1.28.0
python-multipart>=0.0.9



